Improve documentation on the threshold variable
I just realized that the function of threshold, a core component of this program, was left really ambiguous. Basically, any recorded spike with an instFreq value below the specified threshold value will be classified as a cluster (as it sits at either the start or end of a spike cluster).